[IMP] l10n_hr_account_fiskal: l10n_hr_fiskal_user_id
Model on field l10n_hr_fiskal_user_id is changed from res.users to the res.partner so that we have more flexibility (e.g., when an employee without user accounts creates invoices from a POS order, the partner related to the employee will be set as the user who fiscalized invoice).
Added domain to l10n_hr_fiskal_user_id so that only partners related to internal users can be selected.
When an invoice is fiscalized and l10n_hr_fiskal_user_id is not set, then partner related to the invoice_user_id will be used as a default.

[ADD] l10n_hr_pos_employee_fiskal: init

Extended domain on l10n_hr_fiskal_user_id so that partners related to the employees can also be selected.
When an invoice is created from a POS form, a partner related to the employee will be set as a l10n_hr_fiskal_user_id.