[IMP] l10n_hr_account_base: improvements
Added option to show Croatian Payment Means and Fiscal Device on invoice header.
Fiscal Device will be set automatically only if one is available to be selected.

[IMP] l10n_hr_account_fiskal: improvements

Renamed Cash (bills and coins) to Cash (it was too long...)
If the company is not in the tax system, then taxes won't be visible on the printed report.